+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Opt-in `no_nil_response` flags `response = null` and auto-fixes to `response = {}`.
+- Opt-in `no_null_response` flags `response = null` and auto-fixes to `response = {}`.
 
 ## [0.1.0] - 2026-09-08